+++ src/llm/io_processing/gemma4/gemma4_reasoning_parser.cpp
@@ -25,21 +25,36 @@
 #include "../utils.hpp"
 
 namespace ovms {
-void Gemma4ReasoningParser::parse(ParsedOutput& parsedOutput, const std::vector<int64_t>& generatedTokens) {
-    std::string startReasoningTag = getParsingStartTags()[0];
-    std::string endReasoningTag = getParsingEndTag();
-    size_t startPos = parsedOutput.content.find(startReasoningTag);
-    size_t endPos = parsedOutput.content.find(endReasoningTag);
-
-    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
-        size_t reasoningStart = startPos + startReasoningTag.length();
-        std::string reasoningText = parsedOutput.content.substr(reasoningStart, endPos - reasoningStart);
-        parsedOutput.reasoning = reasoningText;
-        // Remove reasoning from content
-        parsedOutput.content.erase(startPos, endPos - startPos + endReasoningTag.length());
+void Gemma4ReasoningParser::skipToken(const std::vector<int64_t>& generatedTokens, size_t& pos, int64_t tokenId) {
+    if (pos < generatedTokens.size() && generatedTokens[pos] == tokenId) {
+        pos++;
     }
 }
 
+void Gemma4ReasoningParser::parse(ParsedOutput& parsedOutput, const std::vector<int64_t>& generatedTokens) {
+    auto startPos = std::string::npos;
+    auto endPos = std::string::npos;
+
+    auto startIt = std::find(generatedTokens.begin(), generatedTokens.end(), channelStartTokenId);
+    auto endIt = std::find(generatedTokens.begin(), generatedTokens.end(), channelEndTokenId);
+
+    if (startIt != generatedTokens.end() && endIt != generatedTokens.end() && startIt < endIt) {
+        startPos = std::distance(generatedTokens.begin(), startIt);
+        endPos = std::distance(generatedTokens.begin(), endIt);
+    }
+
+    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+        skipToken(generatedTokens, startPos, channelStartTokenId);
+        std::string reasoningText = tokenizer.decode(std::vector<int64_t>(generatedTokens.begin() + startPos, generatedTokens.begin() + endPos), ov::genai::skip_special_tokens(true));
+        if (reasoningText.find(reasoningStrIndicator) == 0) {
+            reasoningText = reasoningText.substr(reasoningStrIndicator.size());
+        }
+        parsedOutput.reasoning = reasoningText;
+        // Remove reasoning from content
+        std::string contentWithoutReasoning = tokenizer.decode(std::vector<int64_t>(generatedTokens.begin() + endPos + 1, generatedTokens.end()), ov::genai::skip_special_tokens(true));  // content MUST never appear before reasoning
+        parsedOutput.content = contentWithoutReasoning;
+    }
+}
 std::optional<rapidjson::Document> Gemma4ReasoningParser::parseChunk(const std::string& chunk, ov::genai::GenerationFinishReason finishReason) {
     if (chunk.empty()) {
         SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Received empty chunk for Gemma4ReasoningParser");
