+++ src/llm/visual_language_model/legacy/servable.cpp
@@ -195,8 +195,6 @@ absl::Status VisualLanguageModelLegacyServable::prepareCompleteResponse(std::sha
         executionContext->lastStreamerCallbackOutput.clear();
     }
 
-    executionContext->apiHandler->setPromptTokensUsage(legacyExecutionContext->results.perf_metrics.get_num_input_tokens());
-    executionContext->apiHandler->setCompletionTokensUsage(legacyExecutionContext->results.perf_metrics.get_num_generated_tokens());
     executionContext->response = executionContext->apiHandler->serializeUnaryResponse(legacyExecutionContext->results, completeText);
     SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Complete unary response: {}", executionContext->response);
     return absl::OkStatus();
